requirements.txt
# Core numeric and CV
numpy==1.26.4
opencv-python==4.9.0.80



tflite-runtime==2.12.0
onnxruntime==1.17.3

# Messaging and APIs
paho-mqtt==1.6.1
fastapi==0.111.0
uvicorn[standard]==0.30.0

# Sensors
pynmea2==1.19.0
pyserial==3.5
smbus2==0.4.3
gpiozero==1.6.2

# Config and validation
pydantic==2.7.1
python-dotenv==1.0.1
pyyaml==6.0.1


